fix(models): use int64 for statistics aggregates, tighten devalue override
- statistics.go: change TotalSupporters and TotalInitiatives from int to
  int64 to match the ::bigint SQL cast — pgx v5 maps int8 to int64, not
  int, so scanning into *int would fail at runtime (per copilot[bot])

- package.json: pin pnpm devalue override to ^5.8.1 instead of >=5.8.1
  to avoid accidental major-version selection (per copilot[bot])

+++ backend/internal/domain/models/statistics.go
@@ -8,6 +8,6 @@ package models
 // Values are derived from initiative_ledger_stats which is refreshed by ledger-stats-sync.
 type PlatformStatistics struct {
 	TotalRaisedCents int64 `json:"total_raised_cents"`
-	TotalSupporters  int   `json:"total_supporters"`
-	TotalInitiatives int   `json:"total_initiatives"`
+	TotalSupporters  int64 `json:"total_supporters"`
+	TotalInitiatives int64 `json:"total_initiatives"`
 }
